Refactor to review comments
A review on this code has lead to
the refactoring of the file.

This includes: Removing certain
functionality as width is fixed.
Ierating a test to check more
values.
