+++ src/perm_uses/fixed_hash.rs
@@ -1,12 +1,11 @@
 // Copyright (c) DUSK NETWORK. All rights reserved.
 // Licensed under the MPL 2.0 license. See LICENSE file in the project root for details.”
-//! The `pad` module implements the Sponge's padding algorithm
-use crate::sponge::pad::*;
+//! The `pad` module implements the padding algorithm on the Poseidon hash.
+use super::pad::*;
 
 use dusk_plonk::prelude::*;
 use hades252::strategies::*;
 use hades252::WIDTH;
-
 
 /// Takes in one BlsScalar and outputs 2. 
 /// This function is fixed.
@@ -16,30 +15,15 @@ pub fn two_outputs(message: BlsScalar) -> [BlsScalar; 2] {
     // The value used to pad the words is zero.
     let padder = BlsScalar::zero();
 
-    // The capacity is 
-    let capacity = BlsScalar::one() * BlsScalar::from(2<<64-1) + BlsScalar::one(); 
+       // The capacity is 
+       let capacity = BlsScalar::one() * BlsScalar::from(2<<64-1) + BlsScalar::one(); 
 
     let mut words = pad_fixed_hash(capacity, message, padder);
-    // If the words len is less than the Hades252 permutation `WIDTH` we directly
-    // call the permutation saving useless additions by zero.
-    if words.len() == WIDTH {
-        strategy.perm(&mut words);
-        return [words[1], words[2]];
-    }
-    // If the words len is bigger than the Hades252 permutation `WIDTH` then we
-    // need to collapse the padded limbs. See bottom of pag. 16 of
-    // https://eprint.iacr.org/2019/458.pdf
-    words.chunks(WIDTH).fold(
-        vec![BlsScalar::zero(); WIDTH],
-        |mut inputs, values| {
-            let mut values = values.iter();
-            inputs
-                .iter_mut()
-                .for_each(|input| *input += values.next().unwrap());
-            strategy.perm(&mut inputs);
-            inputs
-        },
-    );
+    // Since we do a fixed_length hash, `words` is always
+    // the size of `WIDTH`. Therefore, we can simply do
+    // the permutation and return the desired results.
+    strategy.perm(&mut words);
+    
     [words[1], words[2]]
 } 
 
@@ -61,11 +45,15 @@ mod tests {
 
     #[test]
     fn same_result() {
+
+        for _i in 0..100 {
         let m = BlsScalar::random(&mut rand::thread_rng()); 
+        
         
         let h = two_outputs(m); 
         let h_1 = two_outputs(m);
-        
+
         assert_eq!(h, h_1);
+        }
     }
 }
