+++ src/perm_uses/mod.rs
@@ -2,3 +2,5 @@
 // Licensed under the MPL 2.0 license. See LICENSE file in the project root for details.”
 
 pub mod fixed_hash;
+
+mod pad;
