+++ src/sponge/pad.rs
@@ -35,25 +35,6 @@ where
     words
 }
 
-// Padding function for a singular scalar input 
-// to provide two outputs from a Poseidon hash
-pub fn pad_fixed_hash <T>(
-    capacity: T,
-    message: T,
-    pad_value: T,
-) -> Vec<T>
-where
-    T: Clone,
-{
-    // For a constant length hash, we always use a slice of width 5.
-    let width = 5;
-    let zero = pad_value;
-    let mut words = vec![zero; width];
-
-    words[0] = capacity;
-    words[1] = message;
-    words
-}
 
 #[cfg(test)]
 mod tests {
